feat(Ingest): Ingest de XML to the database
This part includes the Spring Entity model along with the Repository and the Service which is
transactional. Jackson-Xml added to pom.xml. An ACL is create to transform the xml model to the
domain model. The cron job was replaced with an Event that achieves way better the purpose of
creating the DB scheme.